+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Unreleased
 ---------------------
 
 * Updated EnterpriseCustomerCatalogAdmin save hook to check if a corresponding catalog exists in the enterprise-catalog service. If it does, the save hook will update the existing catalog; otherwise, a new catalog will be created.
+* Added extra logging when syncing Enterprise Catalog data to the Enterprise Catalog Service.
 
 
 [3.2.9] - 2020-05-08